feat: Update SignMessage component with dynamic data
Refactor the SignMessage component to use dynamic data for the image source, title, hostname, and address. This improves the flexibility and reusability of the component. The hardcoded values have been replaced with the corresponding properties from the `signData` object.

chore: Implement SignMessage component for signing messages

Create the SignMessage component to handle the signing of messages. This component provides a user-friendly interface for signing messages and replaces the previous implementation.

fix: Bug fixes

Fix several bugs related to the SignMessage component and light mode. These fixes improve the functionality and user experience of the application.

Create wallet when connect

Implement the functionality to automatically create a wallet when connecting to the application. This simplifies the onboarding process for new users.

Merge pull request #16 from abstraction-hq/5-select-asset-modal

Merge pull request #21 from abstraction-hq/19-implement-handler-for-sign-page

Merge branch '19-implement-handler-for-sign-page' of https://github.com/abstraction-hq/wallet-web into 19-implement-handler-for-sign-page